refactor: move view code into view folder
